Bugfixes
- Backback#addItem would always throw an exception even if the item can be added
- NaiveSolver#solve sorted the items in ascending order while we wanted it to sort items in descending order

+++ src/me/gmz/ISolver.java
@@ -3,5 +3,5 @@ package me.gmz;
 import java.util.ArrayList;
 
 public interface ISolver {
-    Backpack solve(ArrayList<Item> data, int backpackWeight);
+    Backpack solve(ArrayList<Item> data, float backpackWeight);
 }
